Create separate needle models as required
The two tests affected didn't really have anything to do with the
needle model, they were just using it as a convenient item that
had already been created. It seems to me that it's better to not
have them as subclasses of TestNeedle, but rather create the nodes
that they require within the tests themselves. Creating needles
with different names in each test or class avoids the warnings
that were being raised.
The main problem is that we have no way of deleting a model that
has been added, and so all the tests build on each other and are
unable to destroy their fixtures. Maybe it's worth it adding a
function for that.