Avoid segmentation fault
Not fully sure this was happening, but it was introduced when the
QButton was added. Catching this signal seems to stop it happening,
and we can later do something more meaningful to destroy other
things if needed. This didn't happen (or at least wasn't reported)
on Windows, only noticed on Mac.