Set up the nodes for PlusRemote
If the nodes named ScoutScan and liveReconstruction don't exist,
the PlusRemote module will just create new nodes named _1, _2, ...
at every step of the recording and reconstruction. To avoid this,
we can create the nodes with the right names in advance, and then
PlusRemote will just use those and update them as required.