Don't hardcode TiptoTool transforms in PLUS Config
Any transforms specified in the .xml file can't be changed in Slicer.
This means that USTipToUS and StimTipToStim weren't actually being updated
by the pivot calibration.

It just happened that the USTipToUS transform that had been specified was
nearly correct, but the one for the Stim wasn't, which is why there was such
an error with the Stim position.